requirements.txt
tortoise-orm>=0.19.0
asyncpg>=0.27.0
opensimplex>=0.4.0
colorama>=0.4.6
python-dotenv>=0.19.0
